requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
Django>=4.2,<5.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
psutil>=5.9.6
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
requests>=2.31.0
pywebview>=4.4.1
requests>=2.31.0
aiohttp>=3.9.0
telethon>=1.34.0
slack_sdk>=3.27.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
qrcode>=7.4.2
Pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
asyncpg>=0.29.0
